CMR-7529: Refactored code.
