+++ graph-db/serverless/src/utils/cmr/indexPlatform.js
@@ -1,62 +1,9 @@
 import gremlin from 'gremlin'
 
+import { createAcquiredByEdge } from './createAcquiredByEdge'
+import { indexInstrument } from './indexInstrument'
+
 const gremlinStatistics = gremlin.process.statics
-
-/**
- * create the acquiredBy edge between platform/instrument and collection
- * @param {JSON} piVertex the platformInstrument vertex
- * @param {Connection} gremlinConnection a connection to the gremlin server
- * @param {Graph Node} dataset the parent collection vertex in the gremlin server
- * @returns null
- */
-const createAcquiredByEdge = async (piVertex, gremlinConnection, dataset) => {
-  const { value: vertexValue = {} } = piVertex
-  const { id: piId } = vertexValue
-
-  console.log(`PlatformInstrument vertex [${piId}] indexed for collection [${dataset}]`)
-
-  // Create an edge between this platform and its parent collection
-  const platformEdge = await gremlinConnection
-    .V(piId).as('p')
-    .V(dataset)
-    .coalesce(
-      gremlinStatistics.outE('acquiredBy').where(gremlinStatistics.inV().as('p')),
-      gremlinConnection.addE('acquiredBy').to('p')
-    )
-    .next()
-
-  const { value: edgeValue = {} } = platformEdge
-  const { id: edgeId } = edgeValue
-
-  console.log(`acquiredBy edge [${edgeId}] indexed to point to collection [${dataset}]`)
-}
-
-/**
- * Given an Instrument object, Gremlin connection, and associated platform and collection, index instrument and build relationships between platform/instrument and collection
- * @param {JSON} instrument the instrument
- * @param {Connection} gremlinConnection a connection to the gremlin server
- * @param {String} platformName the name of the platform
- * @param {Graph Node} dataset the parent collection vertex in the gremlin server
- * @returns null
- */
-const indexInstrument = async (instrument, gremlinConnection, platformName, dataset) => {
-  const {
-    ShortName: instrumentName
-  } = instrument
-
-  const piVertex = await gremlinConnection
-    .V()
-    .has('platformInstrument', 'platform', platformName)
-    .has('instrument', instrumentName)
-    .fold()
-    .coalesce(
-      gremlinStatistics.unfold(),
-      gremlinConnection.addV('platformInstrument').property('platform', platformName).property('instrument', instrumentName)
-    )
-    .next()
-
-  await createAcquiredByEdge(piVertex, gremlinConnection, dataset)
-}
 
 /**
  * Given a Platform object, Gremlin connection, and associated collection, index platform and instrument and build relationships between platform/instrument and collection
@@ -89,10 +36,14 @@ const indexPlatform = async (platform, gremlinConnection, dataset, conceptId) =>
         )
         .next()
 
-      await createAcquiredByEdge(piVertex, gremlinConnection, dataset)
+      const { value: vertexValue = {} } = piVertex
+      const { id: piId } = vertexValue
+      console.log(`PlatformInstrument vertex [${piId}] indexed for collection [${dataset}]`)
+
+      await createAcquiredByEdge(piId, gremlinConnection, dataset)
     }
   } catch (error) {
-    console.log(`ERROR indexing Platform for concept [${conceptId}] ${JSON.stringify(platform)}: \n Error: ${error}`)
+    console.error(`ERROR indexing Platform for concept [${conceptId}] ${JSON.stringify(platform)}: ${error}`)
   }
 }
